added ignore list for certain version of packages
now 'check' will ignore package if they are on the ignore list. You
can set either a version number for that package or '-' to ignore all
versions. If you set a version number, the package will automatically
be taken off the list once a newer version is available.

new commands:
ignore                      list all packages on the list
ignore <package>            list the version of that package on the list
ignore <package> <version>  adds a new package to the list
removeignore <package>      removes that package from the list